requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.0
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.7.0
idna==3.10
multidict==6.4.4
openai==0.27.7
propcache==0.3.2
psycopg2-binary==2.9.6
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.1
